pyproject.toml
[tool.poetry]
name = "pymemgpt"
version = "0.3.0"
packages = [
    {include = "memgpt"}
]
description = "Teaching LLMs memory management for unbounded context"
authors = [
    "MemGPT Team <hi@memgpt.ai>",
]
license = "Apache License"
readme = "README.md"

[tool.poetry.scripts]
memgpt = "memgpt.main:app"

[tool.poetry.dependencies]
python = "<3.12,>=3.9"
typer = {extras = ["all"], version = "^0.9.0"}
questionary = "^2.0.1"
pytz = "^2023.3.post1"
tqdm = "^4.66.1"
black = { version = "^23.10.1", optional = true }
pytest = { version = "^7.4.4", optional = true }
llama-index = "^0.9.13"
setuptools = "^68.2.2"
datasets = { version = "^2.14.6", optional = true}
prettytable = "^3.9.0"
pgvector = { version = "^0.2.3", optional = true }
huggingface-hub = { version = "0.17.3", optional = true }
transformers = { version = "4.34.1", optional = true }
pre-commit = {version = "^3.5.0", optional = true }
pg8000 = {version = "^1.30.3", optional = true}
torch = {version = ">=2.0.0, !=2.0.1, !=2.1.0", optional = true}
websockets = {version = "^12.0", optional = true}
docstring-parser = "^0.15"
httpx = "^0.25.2"
numpy = "^1.26.2"
demjson3 = "^3.0.6"
tiktoken = "^0.5.1"
python-box = "^7.1.1"
pypdf = "^3.17.1"
pyyaml = "^6.0.1"
chromadb = "^0.4.18"
sqlalchemy-json = "^0.7.0"
fastapi = {version = "^0.104.1", optional = true}
uvicorn = {version = "^0.24.0.post1", optional = true}
pytest-asyncio = {version = "^0.23.2", optional = true}
pydantic = "^2.5.2"
pyautogen = {version = "0.2.0", optional = true}
html2text = "^2020.1.16"
docx2txt = "^0.8"
sqlalchemy = "^2.0.25"
pexpect = {version = "^4.9.0", optional = true}
pyright = {version = "^1.1.347", optional = true}

[tool.poetry.extras]
local = ["torch", "huggingface-hub", "transformers"]
postgres = ["pgvector", "pg8000"]
dev = ["pytest", "pytest-asyncio", "pexpect", "black", "pre-commit", "datasets", "pyright"]
server = ["websockets", "fastapi", "uvicorn"]
autogen = ["pyautogen"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
